requirements.txt
wget
streamlit
torch
Pillow